Move all modules out of the `ordination` subpackage and into the root
