+++ src/geeknn/__init__.py
@@ -1,0 +1,13 @@
+from ._base import RawKNNClassifier
+from .euclidean import EuclideanKNNClassifier
+from .gnn import GNNClassifier
+from .mahalanobis import MahalanobisKNNClassifier
+from .msn import MSNClassifier
+
+__all__ = [
+    "RawKNNClassifier",
+    "EuclideanKNNClassifier",
+    "MahalanobisKNNClassifier",
+    "MSNClassifier",
+    "GNNClassifier",
+]
