+++ tests/test_model_spatial.py
@@ -2,7 +2,7 @@ import ee
 import numpy as np
 import pytest
 
-from geeknn.ordination import (
+from geeknn import (
     EuclideanKNNClassifier,
     GNNClassifier,
     MahalanobisKNNClassifier,
